+++ packages/@runlightyear/salesforce/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/salesforce
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.24.0
+
 ## 0.2.2
 
 ### Patch Changes